Updated strategy.js
Doc & string error to fail the strategy with a message.

Note : I made a similar update in my code, that I was about to commit. In my version, I passed an object containing the key & secret as the second parameter of createProvider, so the provider would be created in strategy.js. In result, no need to import ims-lti in the implementation of createProvider in other files - just saying.